Restructure  - Move part of the auditable responsible for determining auditable attributes to concern.  - small cleanups Note:  - removed a clause excluding on scalar values apart from certain cases. I do not believe we should exclude them by default, but rather handle the case where we are unable to use the value. "hidden" excludes are not desirable  - Removed the protected excludes parameter. just not necessary.
